Reorganize Info and Misc.
Makes the info credits paginated so we can add more text over time with no issues and also moves the Open Appdata/User Folder button to the Misc. menu to avoid cluttering.
