+++ src/pc/djui/djui_panel_options.c
@@ -1,7 +1,3 @@
-#if defined(_WIN32) || defined(_WIN64)
-#include <windows.h>
-#endif
-
 #include "djui.h"
 #include "djui_panel.h"
 #include "djui_panel_player.h"
@@ -15,23 +11,6 @@
 #include "pc/network/network.h"
 #include "pc/utils/misc.h"
 #include "pc/pc_main.h"
-
-static void djui_panel_options_open_user_folder(UNUSED struct DjuiBase* caller) {
-#if defined(_WIN32) || defined(_WIN64)
-    // Windows
-    ShellExecuteA(NULL, "open", sys_user_path(), NULL, NULL, SW_SHOWNORMAL);
-#elif __linux__
-    // Linux
-    char command[512];
-    snprintf(command, sizeof(command), "xdg-open %s", sys_user_path());
-    system(command);
-#elif __APPLE__
-    // macOS
-    char command[512];
-    snprintf(command, sizeof(command), "open %s", sys_user_path());
-    system(command);
-#endif
-}
 
 static void djui_panel_options_back(struct DjuiBase* caller) {
     configfile_save(configfile_name());
@@ -54,11 +33,6 @@ void djui_panel_options_create(struct DjuiBase* caller) {
         djui_button_create(body, DLANG(OPTIONS, DISPLAY), DJUI_BUTTON_STYLE_NORMAL, djui_panel_display_create);
         djui_button_create(body, DLANG(OPTIONS, SOUND), DJUI_BUTTON_STYLE_NORMAL, djui_panel_sound_create);
         djui_button_create(body, DLANG(OPTIONS, MISC), DJUI_BUTTON_STYLE_NORMAL, djui_panel_misc_create);
-#if defined(_WIN32) || defined(_WIN64)
-        djui_button_create(body, DLANG(OPTIONS, APPDATA), DJUI_BUTTON_STYLE_NORMAL, djui_panel_options_open_user_folder);
-#elif __linux__ || __APPLE__ || __MACH__
-        djui_button_create(body, DLANG(OPTIONS, USER_FOLDER), DJUI_BUTTON_STYLE_NORMAL, djui_panel_options_open_user_folder);
-#endif
         djui_button_create(body, DLANG(MENU, BACK), DJUI_BUTTON_STYLE_BACK, djui_panel_options_back);
     }
